FastAPI/requirements.txt
bcrypt==4.0.1
cryptography==39.0.2
fastapi==0.94.1
passlib==1.7.4
pymongo==4.3.3
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pyyaml==6.0
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4